Add single bucket connection type and a file picker input

+++ Apps.AmazonS3/Actions/BucketActions.cs
@@ -1,8 +1,12 @@
 ﻿using Amazon.S3.Model;
+using Apps.AmazonS3.Constants;
+using Apps.AmazonS3.DataSourceHandlers;
 using Apps.AmazonS3.Models.Request;
 using Apps.AmazonS3.Models.Response;
 using Blackbird.Applications.Sdk.Common;
 using Blackbird.Applications.Sdk.Common.Actions;
+using Blackbird.Applications.Sdk.Common.Dynamic;
+using Blackbird.Applications.Sdk.Common.Exceptions;
 using Blackbird.Applications.Sdk.Common.Invocation;
 
 namespace Apps.AmazonS3.Actions;
@@ -10,9 +14,13 @@ namespace Apps.AmazonS3.Actions;
 [ActionList("Buckets")]
 public class BucketActions(InvocationContext invocationContext) : AmazonInvocable(invocationContext)
 {
-    [Action("Create bucket", Description = "Create an S3 bucket.")]
-    public async Task<BucketResponse> CreateBucket([ActionParameter] [Display("Bucket name")] string bucketName)
+    [Action("Create bucket", Description = "Create a new S3 bucket.")]
+    public async Task<BucketResponse> CreateBucket(
+        [ActionParameter] [Display("Bucket name")] string bucketName)
     {
+        if (CurrentConnectionType != ConnectionTypes.AllBuckets)
+            throw new PluginMisconfigurationException($"Currently selected connection supports only '{ConnectedBucket}' bucket. Please, switch to 'All buckets' conection for working with buckets themselves.");
+
         var request = new PutBucketRequest
         {
             BucketName = bucketName,
@@ -24,15 +32,19 @@ public class BucketActions(InvocationContext invocationContext) : AmazonInvocabl
         return new() { BucketName = bucketName };
     }
 
-    [Action("Delete bucket", Description = "Delete an S3 bucket.")]
-    public async Task DeleteBucket([ActionParameter] BucketRequest bucket)
+    [Action("Delete bucket", Description = "Delete an existing S3 bucket.")]
+    public async Task DeleteBucket(
+        [ActionParameter, Display("Bucket name"), DataSource(typeof(BucketDataHandler))] string bucketName)
     {
+        if (CurrentConnectionType != ConnectionTypes.AllBuckets)
+            throw new PluginMisconfigurationException($"Currently selected connection supports only '{ConnectedBucket}' bucket. Please, switch to 'All buckets' conection for working with buckets themselves.");
+
         var request = new DeleteBucketRequest
         {
-            BucketName = bucket.BucketName,
+            BucketName = bucketName,
         };
 
-        var client = await CreateBucketClient(bucket.BucketName);
+        var client = await CreateBucketClient(bucketName);
         await ExecuteAction(() => client.DeleteBucketAsync(request));
     }
 }
